Move globalstyles to Constants
